requirements.txt
pycryptodome==3.15.0
Requests==2.31.0
urllib3==1.25.11
